requirements.txt
pyTelegramBotAPI~=4.13.0
python-dotenv~=0.21.1
telebot~=0.0.4
Pillow~=10.0.0
